Add files via upload
Example of solution of work around of needing different approaches to solve using upright bold Greek letters for vectors and matrices. The easiest solution would be if unicode-math was available in mathjax, but instead, use upgreek and boldsymbol when compiling to html.
The start of the document is creating common commands so enable easy switching between output types; ideally this would just detect output format, but commenting out is easiest approach at the moment.